backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.35
psycopg[binary]>=3.2.3
alembic>=1.13.1

# Authentication and security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Settings management
pydantic-settings>=2.1.0

# Development tools
python-dotenv>=1.0.0
